feat: agregar ubicacion en visitas
- VisitaTableDto: agregar localLatitud, localLongitud, fechaReal, latitudLlegada, longitudLlegada, distanciaMetros
- VisitaDto: agregar localLatitud, localLongitud
- QueryRepository: calcular distancia con formula Haversine (metros)
- Mostrar ubicacion del local y de la visita confirmada

+++ src/main/java/com/cloud_technological/aura_pos/services/LocalService.java
@@ -79,7 +79,7 @@ public class LocalService {
         LocalEntity entity = localRepository.findById(id)
                 .orElseThrow(() -> new GlobalException(HttpStatus.NOT_FOUND, "Local no encontrado"));
 
-        if (!entity.getEmpresa().getId().equals(empresaId.longValue())) {
+        if (!entity.getEmpresa().getId().equals(empresaId)) {
             throw new GlobalException(HttpStatus.NOT_FOUND, "Local no encontrado");
         }
 
@@ -124,7 +124,7 @@ public class LocalService {
         LocalEntity entity = localRepository.findById(id)
                 .orElseThrow(() -> new GlobalException(HttpStatus.NOT_FOUND, "Local no encontrado"));
 
-        if (!entity.getEmpresa().getId().equals(empresaId.longValue())) {
+        if (!entity.getEmpresa().getId().equals(empresaId)) {
             throw new GlobalException(HttpStatus.NOT_FOUND, "Local no encontrado");
         }
 
@@ -179,7 +179,7 @@ public class LocalService {
         LocalEntity entity = localRepository.findById(id)
                 .orElseThrow(() -> new GlobalException(HttpStatus.NOT_FOUND, "Local no encontrado"));
 
-        if (!entity.getEmpresa().getId().equals(empresaId.longValue())) {
+        if (!entity.getEmpresa().getId().equals(empresaId)) {
             throw new GlobalException(HttpStatus.NOT_FOUND, "Local no encontrado");
         }
